chore(boards): cleanup after organizations dropdown
